Added backbone for generic algorithm.
This is the initial commit that introduces actual genetic algorithm logic into
the code. Not tested yet, and definately throws an error. Certain features
incomplete. This commit accomplishes:

Created respective files for genetic algorithm, including a core
chromosome manipulation module and a probability distribution sampling module.

Things still left to do:
1. The sampling module is wholly incomplete.
2. There needs to be a way to access the "best algorithm" so far in the
Learner. This is both to see the final result and the progression.
3. Generic code-cleaning duty.
4. Testing.

It provides a class which generate successive stages of the genetic algorithm,
along with an update mechanism. There's no way to access the final data yet.