Cleanup commit for genetic algorithm code.
This commit takes up on some of the TODOs that were left by the previous one.

Firstly, the ugly non-monadic random generator nonsense from the last commit,
along with the confusing rgen''' names, is gone. Thanks to the monad we now have
the implicit random number generator state propogated throughout the algorithm.

This resulted in some modifications for the Sampler.hs and Utils.hs modules as
well.

Minor documentation improvements.