Wire-in main commit.
This commit adds the actual functionality that is promised in the README
with the optimal circle acutally being found by the genetic algorithm.

The actual use of the functions implied some interface changes, but not
too many.

Currently everything's very inefficient and there are plans to make it better,
but at least we have some working executable.

For display purposes, this involved writing a function that displays a circle.